+++ CHANGELOG.md
@@ -22,3 +22,5 @@ and this project adheres to
 - `Tag` component
 - `FieldsetLegend` component
 - `Fieldset` component
+- `InputButton` component
+- `Button` component